Move bounds checks to separate checker
