+++ src/theory/arith/nl/nl_solver.h
@@ -78,44 +78,6 @@ class NlSolver
                     const std::vector<Node>& xts);
   //-------------------------------------------- lemma schemas
 
-
-  /** check monomial inferred bounds
-   *
-   * Returns a set of valid theory lemmas, based on a
-   * lemma schema that infers new constraints about existing
-   * terms based on mulitplying both sides of an existing
-   * constraint by a term.
-   * For more details, see Section 5 of "Design Theory
-   * Solvers with Extensions" by Reynolds et al., FroCoS 2017,
-   * Figure 5, this is the schema "Multiply".
-   *
-   * Examples:
-   *
-   * x > 0 ^ (y > z + w) => x*y > x*(z+w)
-   * x < 0 ^ (y > z + w) => x*y < x*(z+w)
-   *   ...where (y > z + w) and x*y are a constraint and term
-   *      that occur in the current context.
-   */
-  void checkMonomialInferBounds(
-      const std::vector<Node>& asserts,
-      const std::vector<Node>& false_asserts);
-
-  /** check monomial infer resolution bounds
-   *
-   * Returns a set of valid theory lemmas, based on a
-   * lemma schema which "resolves" upper bounds
-   * of one inequality with lower bounds for another.
-   * This schema is not enabled by default, and can
-   * be enabled by --nl-ext-rbound.
-   *
-   * Examples:
-   *
-   *  ( y>=0 ^ s <= x*z ^ x*y <= t ) => y*s <= z*t
-   *  ...where s <= x*z and x*y <= t are constraints
-   *     that occur in the current context.
-   */
-  void checkMonomialInferResBounds();
-
   //-------------------------------------------- end lemma schemas
  private:
   /** The inference manager that we push conflicts and lemmas to. */
@@ -127,17 +89,6 @@ class NlSolver
   Node d_zero;
   Node d_true;
   Node d_false;
-
-  /** Context-independent database of constraint information */
-  ConstraintDb d_cdb;
-
-  // term -> coeff -> rhs -> ( status, exp, b ),
-  //   where we have that : exp =>  ( coeff * term <status> rhs )
-  //   b is true if degree( term ) >= degree( rhs )
-  std::map<Node, std::map<Node, std::map<Node, Kind> > > d_ci;
-  std::map<Node, std::map<Node, std::map<Node, Node> > > d_ci_exp;
-  std::map<Node, std::map<Node, std::map<Node, bool> > > d_ci_max;
-
 }; /* class NlSolver */
 
 }  // namespace nl
